Build BL32 as static library and export headers
It is expected that, for AArch32 systems, the PSCI library will be
integrated with AArch32 Secure payload. To enable this usecase, it
should be able to export the BL32 component as a library file along
with required headers. This patch adds required support to the Trusted
Firmware build system to create archive and export required headers.

The build option `BUILD_BL32_LIB` is used to determine whether BL32
should be build as a library or not. The build macro
`MAKE_AR_EXPORT` compiles the specified BL sources and archives them.
All the files listed in `EXPORT_FILES` variable is installed in the
`exports` directory created in the platform build folder.

Change-Id: Ic9a017ae592ff5b9bdffba2cafa8fb697e5320de
Signed-off-by: Soby Mathew <soby.mathew@arm.com>