AArch32: Add support for PSCI integration with OP-TEE
This patch adds OP-TEE as one of the candidates for AArch32 Secure
payload integration with PSCI. A new bl32 component `optee` is added
which enables build of PSCI and related platform files as a library.
This library and the associated headers are installed in the `export`
directory which can be then included in the optee build system to
get the BL32 binary.

Currently only the FVP platform is supported for integration with optee.
A sample make commandline to build the PSCI library for OP-TEE integration
is given below:

 # make PLAT=fvp ARCH=aarch32 AARCH32_SP=optee USE_COHERENT_MEM=0 bl32

Change-Id: I17e823453bd073cddb5cc86f71daba65780f7ba1
Signed-off-by: Soby Mathew <soby.mathew@arm.com>