FVP: Support non-identity device memory mapping
The FVP platform code in Trusted Firmware assumes identity physical
to virtual address mapping. Thus may not be true when the PSCI library
is integrated with external AArch32 Secure Payload. To facilitate this,
a new optional platform API `plat_phys_to_virt()` is introduced to
convert physical to virtual address. The default implementation assumes
identity mapping and returns the same address passed into the function.

This patch also does the essential changes to the FVP platform code to
enable it to be integrated with an AArch32 Secure Payload which does
non-identity mapping.

Change-Id: I9cd3f5246e389d59ee3fade0b248ea655f67ce1b
Signed-off-by: Soby Mathew <soby.mathew@arm.com>