Allow overriding of unused stack invalidation in PSCI
The current PSCI stack invalidation code expects the stack size
to be defined by the platform code using the PLATFORM_STACK_SIZE macro.
In cases where PSCI library is integrated with an external AArch32
Secure Payload, it maybe necessary to define the stack size based on
the Secure Payload configuration. This patch refactors the AArch32
unused stack invalidation routine into a new function
`do_unused_stack_invalidate()`, and makes it a weak definition so that
it can be overridden by the Secure Payload.

Change-Id: I2bb398823f165583d341db0ce1dde1e09ace9489
Signed-off-by: Soby Mathew <soby.mathew@arm.com>